+++ src/components/impl/thumbor-image.js
@@ -1,22 +1,35 @@
-import React from "react";
-import emptyWebGif from 'empty-web-gif';
 import omit from '@babel/runtime/helpers/objectWithoutProperties';
+import emptyWebGif from 'empty-web-gif';
 import { func } from 'prop-types';
-import { FocusedImage } from "quintype-js";
+import { FocusedImage } from 'quintype-js';
+import React from 'react';
+import { Helmet } from 'react-helmet';
 import { USED_PARAMS, hashString } from './image-utils';
 
 export function responsiveProps(props) {
   const image = new FocusedImage(props.slug, props.metadata);
+  const { priority = false } = props;
 
   function generatePath(size) {
-    return "//" + props.imageCDN + "/" + image.path(props.aspectRatio, Object.assign({ w: size }, props.imgParams));
+    return '//' + props.imageCDN + '/' + image.path(props.aspectRatio, Object.assign({ w: size }, props.imgParams));
   }
 
-  return {
+  const imageProps = {
     src: generatePath(props.defaultWidth),
-    srcSet: props.widths ? props.widths.map((size) => `${generatePath(size)} ${size}w`).join(",") : undefined,
+    srcSet: props.widths ? props.widths.map((size) => `${generatePath(size)} ${size}w`).join(',') : undefined,
     key: hashString(props.slug),
+    sizes: props?.sizes
+  };
+
+  if (priority) {
+    return {
+      ...imageProps,
+      fetchPriority: 'high'
+    };
   }
+  return {
+    ...imageProps
+  };
 }
 
 export class ThumborImage extends React.Component {
@@ -40,10 +53,22 @@ export class ThumborImage extends React.Component {
   }
   render() {
     const imageProps = this.state.showImage ? responsiveProps(this.props) : { src: emptyWebGif };
-    return React.createElement(this.props.reactTag || "img", Object.assign(imageProps, omit(this.props, USED_PARAMS), {
-      ref: dom => this.dom = dom,
-      className: this.props.className ? `qt-image ${this.props.className}` : 'qt-image'
-    }));
+    return (
+      <>
+        {this.props?.priority && (
+          <Helmet>
+            <link rel="preload" as="image" imagesrcset={imageProps?.srcSet} imagesizes={imageProps?.sizes} />
+          </Helmet>
+        )}
+        {React.createElement(
+          this.props.reactTag || 'img',
+          Object.assign(imageProps, omit(this.props, USED_PARAMS), {
+            ref: (dom) => (this.dom = dom),
+            className: this.props.className ? `qt-image ${this.props.className}` : 'qt-image'
+          })
+        )}
+      </>
+    );
   }
   componentDidMount() {
     this.shouldLazyLoad() && this.context.lazyLoadObserveImage(this.dom, this);
@@ -60,4 +85,4 @@ ThumborImage.contextTypes = {
   lazyLoadObserveImage: func,
   lazyLoadUnobserveImage: func,
   lazyLoadEagerPredicate: func
-}
+};